+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/transaction_controller_test.exs
@@ -250,6 +250,35 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
 
       assert Enum.count(response["token_transfers"]) == 10
     end
+
+    test "single 1155 flattened", %{conn: conn} do
+      token = insert(:token, type: "ERC-1155")
+
+      tx =
+        :transaction
+        |> insert()
+        |> with_block()
+
+      tt =
+        insert(:token_transfer,
+          transaction: tx,
+          block: tx.block,
+          block_number: tx.block_number,
+          token_contract_address: token.contract_address,
+          token_ids: [1],
+          amounts: [2],
+          amount: nil
+        )
+
+      request = get(conn, "/api/v2/transactions/" <> to_string(tx.hash))
+
+      assert response = json_response(request, 200)
+      compare_item(tx, response)
+
+      assert Enum.count(response["token_transfers"]) == 1
+      assert is_map(Enum.at(response["token_transfers"], 0)["total"])
+      assert compare_item(%TokenTransfer{tt | amount: 2}, Enum.at(response["token_transfers"], 0))
+    end
   end
 
   describe "/transactions/{tx_hash}/internal-transactions" do
@@ -1028,6 +1057,11 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
     compare_item(Enum.at(txs, 0), Enum.at(second_page_resp["items"], 0), wl_names)
   end
 
+  # with the current implementation no transfers should come with list in totals
+  defp check_total(%Token{type: nft}, json, _token_transfer) when nft in ["ERC-721", "ERC-1155"] and is_list(json) do
+    false
+  end
+
   defp check_total(%Token{type: nft}, json, token_transfer) when nft in ["ERC-1155"] do
     json["token_id"] in Enum.map(token_transfer.token_ids, fn x -> to_string(x) end) and
       json["value"] == to_string(token_transfer.amount)
@@ -1035,11 +1069,6 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
 
   defp check_total(%Token{type: nft}, json, token_transfer) when nft in ["ERC-721"] do
     json["token_id"] in Enum.map(token_transfer.token_ids, fn x -> to_string(x) end)
-  end
-
-  # with the current implementation no transfers should come with list in totals
-  defp check_total(%Token{type: nft}, json, _token_transfer) when nft in ["ERC-721", "ERC-1155"] and is_list(json) do
-    false
   end
 
   defp check_total(_, _, _), do: true