+++ CHANGELOG.md
@@ -12,8 +12,11 @@
 
 ### Fixes
 
+- [#7635](https://github.com/blockscout/blockscout/pull/7635) - Fix single 1155 transfer displaying
+- [#7629](https://github.com/blockscout/blockscout/pull/7629) - Fix NFT fetcher
+- [#7614](https://github.com/blockscout/blockscout/pull/7614) - API and smart-contracts fixes and improvements
 - [#7611](https://github.com/blockscout/blockscout/pull/7611) - Fix tokens pagination
-- [#7566](https://github.com/blockscout/blockscout/pull/7566) - Account: check composed email beofre sending
+- [#7566](https://github.com/blockscout/blockscout/pull/7566) - Account: check composed email before sending
 - [#7564](https://github.com/blockscout/blockscout/pull/7564) - Return contract type in address view
 - [#7562](https://github.com/blockscout/blockscout/pull/7562) - Remove fallback from Read methods
 - [#7537](https://github.com/blockscout/blockscout/pull/7537), [#7553](https://github.com/blockscout/blockscout/pull/7553) - Withdrawals fixes and improvements